add dpp -> js document error conversion
