+++ js-binding/src/data_contract/data_contract.rs
@@ -1,5 +1,4 @@
-use crate::errors;
-use crate::errors::RustConversionError;
+use crate::errors::{from_dpp_err, RustConversionError};
 use crate::{bail_js, with_js_error};
 use dpp::data_contract::DataContract;
 use dpp::util::string_encoding::Encoding;
@@ -119,7 +118,7 @@ impl DataContractWasm {
 
     #[wasm_bindgen(js_name=getDocumentSchema)]
     pub fn get_document_schema(&mut self, doc_type: &str) -> Result<JsValue, JsValue> {
-        let doc_schema = self.0.get_document_schema(doc_type).map_err(errors::from)?;
+        let doc_schema = self.0.get_document_schema(doc_type).map_err(from_dpp_err)?;
         with_js_error!(JsValue::from_serde(doc_schema))
     }
 
@@ -129,7 +128,7 @@ impl DataContractWasm {
             &self
                 .0
                 .get_document_schema_ref(doc_type)
-                .map_err(errors::from)?
+                .map_err(from_dpp_err)?
         ))
     }
 
@@ -206,29 +205,29 @@ impl DataContractWasm {
 
     #[wasm_bindgen(js_name=toBuffer)]
     pub fn to_buffer(&self) -> Result<Vec<u8>, JsValue> {
-        self.0.to_buffer().map_err(errors::from)
+        self.0.to_buffer().map_err(from_dpp_err)
     }
 
     #[wasm_bindgen(js_name=hash)]
     pub fn hash(&self) -> Result<Vec<u8>, JsValue> {
-        self.0.hash().map_err(errors::from)
+        self.0.hash().map_err(from_dpp_err)
     }
 
     #[wasm_bindgen(js_name=from)]
     pub fn from(v: JsValue) -> Result<DataContractWasm, JsValue> {
         let json_contract: Value = with_js_error!(v.into_serde())?;
         Ok(DataContract::try_from(json_contract)
-            .map_err(errors::from)?
+            .map_err(from_dpp_err)?
             .into())
     }
 
     #[wasm_bindgen(js_name=from_buffer)]
     pub fn from_buffer(b: Vec<u8>) -> Result<DataContractWasm, JsValue> {
-        Ok(DataContract::from_buffer(b).map_err(errors::from)?.into())
+        Ok(DataContract::from_buffer(b).map_err(from_dpp_err)?.into())
     }
 
     #[wasm_bindgen(js_name=from_string)]
     pub fn from_string(v: &str) -> Result<DataContractWasm, JsValue> {
-        Ok(DataContract::try_from(v).map_err(errors::from)?.into())
+        Ok(DataContract::try_from(v).map_err(from_dpp_err)?.into())
     }
 }
