+++ js-binding/src/mocks.rs
@@ -11,6 +11,14 @@ impl DocumentTransitionWasm {
     }
 }
 
-pub struct ConsensusError {
-    // how to convert all these errors into a logical structure
+impl From<mocks::DocumentTransition> for DocumentTransitionWasm {
+    fn from(v: mocks::DocumentTransition) -> Self {
+        DocumentTransitionWasm(v)
+    }
+}
+
+pub struct ConsensusError {}
+
+pub fn from_consensus_to_js_error(_: mocks::ConsensusError) -> JsValue {
+    unimplemented!()
 }
